refactor: vendor color palette, remove ml4t-style dependency
Move COLORS dict into visualization/_colors.py instead of
depending on unpublished ml4t-style package. Fixes CI failure
where ml4t-style was unavailable via PyPI.

17 import sites updated, pyproject.toml dependency removed.
