+++ src/ml4t/diagnostic/evaluation/visualization.py
@@ -20,7 +20,7 @@ if TYPE_CHECKING:
     from numpy.typing import NDArray
 
 # Colors from canonical ML4T palette + local semantic aliases
-from ml4t.style import COLORS as _ML4T_COLORS
+from ml4t.diagnostic.visualization._colors import COLORS as _ML4T_COLORS
 
 COLORS = {
     "positive": _ML4T_COLORS["positive"],
