+++ src/ml4t/diagnostic/visualization/backtest/cost_attribution.py
@@ -16,9 +16,9 @@ from typing import TYPE_CHECKING, Literal
 
 import numpy as np
 import plotly.graph_objects as go
-from ml4t.style import COLORS as _ML4T_COLORS
 from plotly.subplots import make_subplots
 
+from ml4t.diagnostic.visualization._colors import COLORS as _ML4T_COLORS
 from ml4t.diagnostic.visualization.core import get_theme_config
 
 if TYPE_CHECKING:
