+++ src/ml4t/diagnostic/visualization/backtest/statistical_validity.py
@@ -16,8 +16,8 @@ from typing import TYPE_CHECKING, Any, Literal
 
 import numpy as np
 import plotly.graph_objects as go
-from ml4t.style import COLORS as _ML4T_COLORS
 
+from ml4t.diagnostic.visualization._colors import COLORS as _ML4T_COLORS
 from ml4t.diagnostic.visualization.core import get_theme_config
 
 if TYPE_CHECKING:
