+++ src/ml4t/diagnostic/visualization/barrier_plots.py
@@ -18,8 +18,8 @@ from typing import TYPE_CHECKING
 
 import numpy as np
 import plotly.graph_objects as go
-from ml4t.style import COLORS as _ML4T_COLORS
 
+from ml4t.diagnostic.visualization._colors import COLORS as _ML4T_COLORS
 from ml4t.diagnostic.visualization.core import (
     create_base_figure,
     get_colorscale,
